# docs/swagger.yaml
openapi: 3.1.0
info:
  title: MyTrello API
  version: 1.0.0
  description: API documentation for MyTrello project
...
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/preferences:
    get:
      tags:
        - users
      summary: Get user preferences
      description: Retrieve the preferences for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Preferences retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
        "401":
          $ref: "#/components/responses/Error401"
        "404":
          $ref: "#/components/responses/Error404"
        "500":
          $ref: "#/components/responses/Error500"
    patch:
      tags:
        - users
      summary: Update user preferences
      description: Update one or more preferences for the authenticated user. Only provided fields will be updated.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                localization:
                  type: string
                  description: Localization setting (e.g., "en/US", "fr/FR")
                  example: "en/US"
                theme:
                  type: string
                  enum: [dark, light, system]
                  description: UI theme preference
                  example: "system"
                showAssignedCardsInHomepage:
                  type: boolean
                  description: Whether to show assigned cards on the homepage
                  example: true
      responses:
        "200":
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "500":
          $ref: "#/components/responses/Error500"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      description: Standard error payload
      required:
        - error
      properties:
        error:
          type: string
          description: Human readable error description
          example: Error XXX because of YYY
...
    UserPreferences:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        localization:
          type: string
          description: Localization setting
          example: "en/US"
        theme:
          type: string
          enum: [dark, light, system]
          description: UI theme preference
          example: "system"
        showAssignedCardsInHomepage:
          type: boolean
          description: Whether to show assigned cards on the homepage
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Error500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"